backend/requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
cachetools==6.2.4
certifi==2025.11.12
channels==4.3.2
channels_redis==4.3.0
chardet==3.0.4
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.8
django-allauth==65.13.1
djangorestframework==3.16.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.31.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
msgpack==1.1.2
numpy==2.2.6
oauthlib==3.3.1
pillow==12.0.0
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
pytz==2025.2
qrcode==8.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.11.0
gunicorn==21.2.0
